fixed lint warning

+++ src/scip/presol_symmetry.c
@@ -384,7 +384,7 @@ SCIP_RETCODE collectCoefficients(
       matrixdata->rhscoef[nrhscoef] = rhs;
 
       /* if we deal with special constraints */
-      if ( (int) rhssense >= SYM_SENSE_XOR )
+      if ( (int) rhssense >= (int) SYM_SENSE_XOR )
          matrixdata->rhssense[nrhscoef] = rhssense;
       else
          matrixdata->rhssense[nrhscoef] = SYM_SENSE_EQUATION;
